Refactor Helpers Classes
